+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### Changes:
+- Fix: Align iOS and Android willFit parameter for recommended text
+- Feature: Added Sentry logger
+- Feature: Web view logs handler
+- Bugfix: Load product on switch store changed
+- Codefix: Re-implementation session key requests
+- Codefix: Authorize web view fixed
+
 ### 2.12.21
 - Fix: Compose VirtusizeInPageMini VirtusizeInPageStandard update fix